moves reader to a separate module
